+++ src/site/markdown/ReleaseNotes.md
@@ -2,6 +2,12 @@
 
 These are the release notes for the Java library for mapcodes.
 
+### 2.4.0
+
+* Added scripts for Korean (Choson'gul/Hangul), Burmese, Khmer, Sinhalese, Thaana (Maldivan), Chinese (Zhuyin, Bopomofo).
+
+* Renamed constant `MALAY` to `MALAYALAM`.
+
 ### 2.3.2
 
 * Updated dependencies.